+++ src/molecules/ElectricityWeekChart/fragments.js
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const CONSUMPTION = gql`
-	fragment consumption on ElectricityExchange {
+	fragment ElectricityExchange on ElectricityExchange {
 		received
 		delivered
 		period {