+++ src/molecules/GasReceivedWeekChart/fragments.js
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const CONSUMPTION = gql`
-	fragment consumption on GasExchange {
+	fragment GasExchange on GasExchange {
 		received
 		period {
 			start