+++ src/molecules/WaterReceivedWeekChart/fragments.js
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const CONSUMPTION = gql`
-	fragment consumption on WaterConsumption {
+	fragment WaterConsumption on WaterConsumption {
 		received
 		period {
 			start