+++ src/pages/Portal/Portal.js
@@ -13,18 +13,9 @@ import Widget from '../../molecules/Widget/Widget';
 
 import BoltArrowUpIcon from '../../atoms/BoltArrowUpIcon/BoltArrowUpIcon';
 import BoltArrowDownIcon from '../../atoms/BoltArrowDownIcon/BoltArrowDownIcon';
-import WaterIcon from '../../atoms/WaterIcon/WaterIcon';
-import FireIcon from '../../atoms/FireIcon/FireIcon';
 
 import CurrentElectricityReceived from '../../molecules/CurrentElectricityReceived/CurrentElectricityReceived';
 import CurrentElectricityDelivered from '../../molecules/CurrentElectricityDelivered/CurrentElectricityDelivered';
-import ElectricityReceived from '../../molecules/ElectricityReceived/ElectricityReceived';
-import ElectricityDelivered from '../../molecules/ElectricityDelivered/ElectricityDelivered';
-import GasUsage from '../../molecules/GasUsage/GasUsage';
-import WaterUsage from '../../molecules/WaterUsage/WaterUsage';
-
-import ElectricityMonthChart from '../../molecules/ElectricityMonthChart/ElectricityMonthChart';
-import GasMonthChart from '../../molecules/GasMonthChart/GasMonthChart';
 
 export default function Portal() {
 	return (
@@ -39,24 +30,6 @@ export default function Portal() {
 				</Widget>
 				<Widget title="Delivering" name="current-electricity-delivered" icon={<BoltArrowUpIcon />}>
 					<CurrentElectricityDelivered />
-				</Widget>
-				<Widget title="Today" name="electricity-usage" icon={<BoltArrowDownIcon />}>
-					<ElectricityReceived />
-				</Widget>
-				<Widget title="Today" name="electricity-delivered" icon={<BoltArrowUpIcon />}>
-					<ElectricityDelivered />
-				</Widget>
-				<Widget title="Today" name="gas-usage" icon={<FireIcon />}>
-					<GasUsage />
-				</Widget>
-				<Widget title="Today" name="water-usage" icon={<WaterIcon />}>
-					<WaterUsage />
-				</Widget>
-				<Widget title="Month" name="electrical-usage-by-month" icon={<BoltArrowDownIcon />}>
-					<ElectricityMonthChart />
-				</Widget>
-				<Widget title="Month" name="gas-usage-by-month" icon={<FireIcon />}>
-					<GasMonthChart />
 				</Widget>
 			</WidgetGrid>
 		</LoggedOut>