+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.12.0] - 2021-05-23
+### Added
+- Electricity received to month chart.
+- Last updated time for current electricity delivered and received widgets.
+
+### Removed
+- All widgets (except actual receiving and delivering widgets) on portal page (home page when logged out).
+
 ## [1.11.0] - 2021-05-13
 ### Added
 - Current water meter reading on settings page.